requirements.txt
google-cloud-storage>=2.7.0
requests>=2.28.1